fix(coverage): restore server coverage to 99.5% for Coveralls gate (#316)
The Coveralls check compares server coverage against main's 100% baseline and
fails on a >~1% drop. The migration dropped it to 97.3% because the E2E-login
block (live network calls, untestable in unit tests) was inline in
auth-routes.ts and counted against the file. Three fixes:

- Extracted the E2E-login handler into src/hono/e2e-auth-routes.ts (its own
  module, excluded from the unit-coverage gate — same as the former
  src/routes/e2e-auth-routes.ts was under src/routes/**). auth-routes.ts now
  only mounts it when E2E_TESTING && NODE_ENV !== production.
- Added tests for the defense-in-depth isValidHandle / isValidDid guards
  (reachable by sending handles/DIDs that satisfy the Zod schema but fail the
  AT Protocol format check — "invalid_handle" and "not-a-did").
- Removed the dead "missing required fields" guard in /messages/respond (Zod's
  min(1) on tid/recipient/response makes it unreachable).

Coverage now 99.49% lines / 99.29% functions (was 97.3%). Remaining gaps are
session-fallback edge branches and a fire-and-forget .catch — all hard to
reach without contrived setups.

Note on bun run test:coverage exit code: Bun 1.3.14's coverageThreshold
enforcement is buggy — it exits 1 for ANY coverage below 100% even when above
the configured threshold (verified: threshold=0 with 98% still fails). The CI
Server Tests step uses continue-on-error, so this is non-blocking; the real
gate is Coveralls, which this commit targets.

332 pass, 0 fail. tsc clean.

Refs #316.
