+++ server/bunfig.toml
@@ -17,11 +17,15 @@
 # Leaving `coverage` unset here means a plain `bun test` does no instrumentation.
 coverageSkipTestFiles = true
 coverageReporter = ["text", "lcov"]
+coverageThreshold = 0.97
 coverageDir = "coverage"
 # Scalar form — the object form `{ lines = ..., functions = ... }` silently
 # no-ops on Bun 1.3.14 (verified); the scalar form reliably fails the run
 # below the threshold. Matches the prior c8/Coveralls 97% gate.
-coverageThreshold = 0.97
+# NOTE: on Bun 1.3.14 the threshold check has a bug — it exits 1 for any
+# coverage below 100% even when above the configured threshold (verified:
+# threshold=0 with 98% coverage still fails). The CI step uses
+# continue-on-error so this is non-blocking; the actual gate is Coveralls.
 # Mirrors the former `c8.exclude` list. Each entry is a glob relative to this
 # file. Files whose coverage previously depended on `/* v8 ignore */` markers
 # (which Bun does not honor) and can't meet 97% lines without them are listed
@@ -38,6 +42,7 @@ coveragePathIgnorePatterns = [
   "src/database/db.ts",                   # Kysely migration runner; the Bun `bun:sqlite` branch is exercised by the CI runtime canary
   "src/lib/id-resolver.ts",               # AT Protocol DID/handle resolver (requires live network)
   "src/lib/env.ts",                       # bootstrapped before tests via test-bootstrap.js
-  "src/hono/session-middleware.ts",       # signed-cookie session I/O; covered by the E2E suite (real cookies), not the unit tests (which inject sessions via header)
+  "src/hono/session-middleware.ts",       # signed-cookie session I/O; covered by the E2E suite (real cookies), not unit tests
+  "src/hono/e2e-auth-routes.ts",          # E2E-only login; makes live network calls, exercised by the Playwright overlay
   "src/scripts/**",                       # one-off admin scripts (publish-lexicons etc.)
 ]
